+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - A fallback value processing from the \<paint\> type.
 - `has_attributes`, `remove`, `is_basic_shape` methods to the `Node`.
 - `is_graphical_event`, `is_conditional_processing`, `is_core`, `is_fill`, `is_stroke` methods to the `Attribute`.
+- `types::path::Segment` struct which is used instead of one from `libsvgparser`.
+- `to_absolute` and `to_relative` methods to the `types::path::Path`.
 
 ### Changed
 - Moved back from `dtoa` to the std implementation.